ov004: the 0x020b04e8 run is TWO TUs, and the pragma proves it. dScMgBase_c 8/8
Correcting my own reasoning from two commits ago. I argued the 50-function
contiguous run 0x020b04e8..0x020b2adc was one TU because one object's .text is
contiguous. Contiguity is necessary but NOT sufficient -- ADJACENT objects are
contiguous too -- so that argument only bounded the run from above.

The pragma settles it from the other side. Three members disagree about
opt_strength_reduction and it cannot be satisfied twice in one file:

  _ZN11dScMgBase_c14BeforeBehaviorEv  0x020b0620  needs it OFF
  func_ov004_020b0cac                 0x020b0cac  needs it ON
  func_ov004_020b14f0                 0x020b14f0  needs it ON

Measured three ways, and the third measurement pins down the semantics:

  no pragma (SR on)                 47/49   BeforeBehavior differs
  file-global `off`                 46/49   0b0cac + 0b14f0 differ
  `off` before + `on` after the
    one definition, bracketing it   47/49   BeforeBehavior differs

The third is the informative one. Bracketing a single definition does NOT scope
the pragma to it: everything compiled with SR ON, exactly as if the bracket were
absent. So opt_strength_reduction obeys the same rule the skill documents for
opt_propagation / optimize_for_size -- the LAST state set anywhere applies to the
whole file, position is irrelevant. There is no source shape that gives
BeforeBehavior a fixed base plus `add r0, r5, r7, lsl#5` while leaving the other
two strength-reduced, because that IS the pragma's effect.

Splitting at 0x020b0a38 -- exactly where the `_ZN11dScMgBase_c*` cluster ends and
the `func_ov004_*` run begins -- satisfies both states and both sub-TUs verify
better than the merged one did:

  ov004/dScMgBase_c      0x020b04e8..0x020b0a38   8 members, pragma off
      byte comparison   : 8/8 MATCH
      objisolate check  : clean
      reloc-destinations: clean
      -> TEXT-VERIFIED

  ov004/unit020b0a38     0x020b0a38..0x020b2adc  41 members, no pragma
      byte comparison   : 40/41 MATCH
      -> still shadow; func_ov004_020b0b1c differs, and 9 unlicensed .data

48/49 members now reproduce, against 47/49 merged.

TU-A came out clean of unlicensed .data, which corroborates the split: the key
function is the destructor (declared first in include/dScMgBase_c.h), the
destructor group lives in TU-B, and so _ZTV/_ZTI/_ZTS are emitted by TU-B alone.
TU-A overrides eight other virtuals and none of them is the key function.
