+++ config/tu_manifest.json
@@ -1944,10 +1944,10 @@
       "module": "ov004",
       "source": "src_tu/actors/dScMgBase_c.cpp",
       "promoted_source": "src/actors/dScMgBase_c.cpp",
-      "status": "shadow",
+      "status": "text-verified",
       "boundary_confidence": "medium",
       "boundary_evidence": [
-        "contiguous linker run: 0x20b04e8..0x20b2adc, 49 function(s), from build/tu_map.json (tools/tu_map.py)",
+        "contiguous linker run: 0x20b04e8..0x20b0a38, 8 function(s). This is a SUB-RANGE of build/tu_map.json's ov004 unit 9 (0x020b04e8..0x020b2adc), split because #pragma opt_strength_reduction is file-global and the ROM needs it OFF for 0x020b0620 and ON for 0x020b0cac / 0x020b14f0 -- see this file's own header comment",
         "class label(s): dScMgBase_c",
         "module sinit corroboration: 4 sinit(s) / 4 .ctor entries, sinit_vs_tu=ok, corroborated=False (module-wide, NOT narrowed to this one TU -- see notes/tu-reconstruction-pilot-report.md sec 2 for the by-hand narrowing step)"
       ],
@@ -1955,7 +1955,7 @@
         {
           "name": ".text",
           "start": "0x020b04e8",
-          "end": "0x020b2adc"
+          "end": "0x020b0a38"
         }
       ],
       "functions": [
@@ -2014,7 +2014,50 @@
           "size": "0x00000108",
           "legacy_source": "src/_ZN11dScMgBase_c19BeforeInitResourcesEv.cpp",
           "ordinal": 15
-        },
+        }
+      ],
+      "data": [],
+      "bss": [],
+      "notes": [
+        "Generated by tools/tubuild.py create. See the file's own header comment for what was and was not reconciled, and notes/translation-unit-reconstruction-plan.md section 7.3 for what this generator does and does not attempt to resolve automatically."
+      ],
+      "verification": {
+        "round": "tools/tubuild.py verify -- text-only, whole shadow TU compiled once",
+        "toolchain": "tools/mwccarm/2004/b56/mwccarm.exe",
+        "flags": "-O4,p -enum int -lang c++ -char signed -interworking -proc arm946e -gccext,on -msgstyle gcc -Cpp_exceptions off",
+        "object": "build/tu/ov004-dScMgBase_c/dScMgBase_c.o (gitignored)",
+        "comparison": "tools/match.py extract_func+compare (bytes/relocs) AND tools/objisolate.py plan() (relocation type/addend) AND tools/reloc_audit.py check_destinations() (relocation target identity)",
+        "functions_matched": 8,
+        "functions_declared": 8,
+        "criteria": {
+          "every_declared_function_defined": "PASS",
+          "every_declared_function_bytes_match": "PASS",
+          "declared_function_set_equals_defined_function_set": "PASS",
+          "functions_occur_in_expected_order": "PASS",
+          "relocation_destinations_verified": "PASS"
+        }
+      }
+    },
+    {
+      "id": "ov004/unit020b0a38",
+      "module": "ov004",
+      "source": "src_tu/actors/unit_ov004_020b0a38.cpp",
+      "promoted_source": "src/actors/unit020b0a38.cpp",
+      "status": "shadow",
+      "boundary_confidence": "medium",
+      "boundary_evidence": [
+        "contiguous linker run: 0x20b0a38..0x20b2adc, 41 function(s). This is a SUB-RANGE of build/tu_map.json's ov004 unit 9 (0x020b04e8..0x020b2adc), split because #pragma opt_strength_reduction is file-global and the ROM needs it OFF for 0x020b0620 and ON for 0x020b0cac / 0x020b14f0 -- see this file's own header comment",
+        "class label(s): dScMgBase_c",
+        "module sinit corroboration: 4 sinit(s) / 4 .ctor entries, sinit_vs_tu=ok, corroborated=False (module-wide, NOT narrowed to this one TU -- see notes/tu-reconstruction-pilot-report.md sec 2 for the by-hand narrowing step)"
+      ],
+      "sections": [
+        {
+          "name": ".text",
+          "start": "0x020b0a38",
+          "end": "0x020b2adc"
+        }
+      ],
+      "functions": [
         {
           "symbol": "func_ov004_020b0a38",
           "address": "0x020b0a38",
@@ -2308,7 +2351,6 @@
       "notes": [
         "Generated by tools/tubuild.py create. See the file's own header comment for what was and was not reconciled, and notes/translation-unit-reconstruction-plan.md section 7.3 for what this generator does and does not attempt to resolve automatically.",
         "tubuild create warning: CONFLICT: extern declaration 'data_ov004_020beb68' differs between the legacy file that used to hold func_ov004_020b0a54 and the one that held func_ov004_020b0b1c; kept the first, the other is commented out for review",
-        "tubuild create warning: CONFLICT: extern declaration 'FreeGfxSlotsById' differs between the legacy file that used to hold _ZN11dScMgBase_c21AfterCleanupResourcesEj and the one that held func_ov004_020b0d30; kept the first, the other is commented out for review",
         "tubuild create warning: CONFLICT: extern declaration 'GetGameLanguage' differs between the legacy file that used to hold func_ov004_020b0d8c and the one that held func_ov004_020b0e84; kept the first, the other is commented out for review",
         "tubuild create warning: CONFLICT: extern declaration 'GetGameLanguage' differs between the legacy file that used to hold func_ov004_020b0d8c and the one that held func_ov004_020b1710; kept the first, the other is commented out for review",
         "tubuild create warning: CONFLICT: extern declaration 'func_ov004_020b1ea4' differs between the legacy file that used to hold func_ov004_020b0e84 and the one that held func_ov004_020b1710; kept the first, the other is commented out for review",
@@ -2325,6 +2367,7 @@
         "tubuild create warning: CONFLICT: extern declaration 'func_ov004_020b1d60' differs between the legacy file that used to hold func_ov004_020b1de8 and the one that held func_ov004_020b2444; kept the first, the other is commented out for review",
         "tubuild create warning: CONFLICT: extern declaration 'GetGameLanguage' differs between the legacy file that used to hold func_ov004_020b0d8c and the one that held func_ov004_020b2574; kept the first, the other is commented out for review",
         "tubuild create warning: CONFLICT: extern declaration 'DrawOamSprite' differs between the legacy file that used to hold func_ov004_020b1e44 and the one that held func_ov004_020b2574; kept the first, the other is commented out for review",
+        "tubuild create warning: CONFLICT: extern declaration 'FreeGfxSlotsById' differs between the legacy file that used to hold func_ov004_020b0d30 and the one that held func_ov004_020b265c; kept the first, the other is commented out for review",
         "tubuild create warning: CONFLICT: extern declaration 'GetGameLanguage' differs between the legacy file that used to hold func_ov004_020b0d8c and the one that held func_ov004_020b27f4; kept the first, the other is commented out for review",
         "tubuild create warning: CONFLICT: extern declaration 'GetGameLanguage' differs between the legacy file that used to hold func_ov004_020b0d8c and the one that held func_ov004_020b2880; kept the first, the other is commented out for review"
       ],
@@ -2332,10 +2375,10 @@
         "round": "tools/tubuild.py verify -- text-only, whole shadow TU compiled once",
         "toolchain": "tools/mwccarm/2004/b56/mwccarm.exe",
         "flags": "-O4,p -enum int -lang c++ -char signed -interworking -proc arm946e -gccext,on -msgstyle gcc -Cpp_exceptions off",
-        "object": "build/tu/ov004-dScMgBase_c/dScMgBase_c.o (gitignored)",
+        "object": "build/tu/ov004-unit020b0a38/unit_ov004_020b0a38.o (gitignored)",
         "comparison": "tools/match.py extract_func+compare (bytes/relocs) AND tools/objisolate.py plan() (relocation type/addend) AND tools/reloc_audit.py check_destinations() (relocation target identity)",
-        "functions_matched": 47,
-        "functions_declared": 49,
+        "functions_matched": 40,
+        "functions_declared": 41,
         "criteria": {
           "every_declared_function_defined": "PASS",
           "every_declared_function_bytes_match": "FAIL",
