+++ src_tu/actors/unit_ov004_020b0a38.cpp
@@ -1,0 +1,1301 @@
+//cpp
+/* SHADOW translation unit -- generated by tools/tubuild.py create.
+ * ov004/unit020b0a38  (41 function(s))
+ *
+ * NOT ENROLLED, NOT CANONICAL. This file contributes nothing to the ROM
+ * build. It is a STARTING POINT (plan sec 7.3): local shadow declarations
+ * below were carried verbatim from the legacy files, not reconciled against
+ * real project headers -- that judgement call is left to a human/LLM review,
+ * the way pilot #1 reconciled PoleLift::Render and ::CleanupResources by hand
+ * (see notes/tu-reconstruction-pilot-report.md sec 5.2).
+ *
+ * FUNCTION ORDER IS DELIBERATELY THE REVERSE OF THE ROM'S -- mwccarm 2004/b56
+ * emits one .text section per function, in the REVERSE of source order, so
+ * the highest-address ROM function is written FIRST here. Do not reorder;
+ * see notes/tu-reconstruction-pilot-report.md sec 3 for the one documented
+ * exception (a destructor's D0/D1/D2 group has compiler-chosen order).
+ *
+ * Assembled from these legacy one-function sources (ROM address order):
+ *   [16] 0x020b0a38  src/func_ov004_020b0a38.c
+ *   [17] 0x020b0a54  src/func_ov004_020b0a54.c
+ *   [18] 0x020b0aa0  src/FreeGfxSlotsById.c
+ *   [19] 0x020b0b1c  src/func_ov004_020b0b1c.c
+ *   [20] 0x020b0cac  src/func_ov004_020b0cac.c
+ *   [21] 0x020b0d30  src/func_ov004_020b0d30.c
+ *   [22] 0x020b0d8c  src/func_ov004_020b0d8c.c
+ *   [23] 0x020b0de0  src/func_ov004_020b0de0.c
+ *   [24] 0x020b0e84  src/func_ov004_020b0e84.cpp
+ *   [25] 0x020b14f0  src/func_ov004_020b14f0.c
+ *   [26] 0x020b1710  src/func_ov004_020b1710.c
+ *   [27] 0x020b19f0  src/func_ov004_020b19f0.c
+ *   [28] 0x020b1a5c  src/func_ov004_020b1a5c.c
+ *   [29] 0x020b1aec  src/func_ov004_020b1aec.c
+ *   [30] 0x020b1b08  src/func_ov004_020b1b08.c
+ *   [31] 0x020b1b40  src/func_ov004_020b1b40.c
+ *   [32] 0x020b1b78  src/func_ov004_020b1b78.c
+ *   [33] 0x020b1ba0  src/func_ov004_020b1ba0.c
+ *   [34] 0x020b1bc8  src/func_ov004_020b1bc8.c
+ *   [35] 0x020b1c68  src/func_ov004_020b1c68.c
+ *   [36] 0x020b1cf0  src/func_ov004_020b1cf0.c
+ *   [37] 0x020b1d60  src/func_ov004_020b1d60.c
+ *   [38] 0x020b1de8  src/func_ov004_020b1de8.c
+ *   [39] 0x020b1e34  src/func_ov004_020b1e34.c
+ *   [40] 0x020b1e44  src/func_ov004_020b1e44.c
+ *   [41] 0x020b1ea4  src/func_ov004_020b1ea4.c
+ *   [43] 0x020b2444  src/func_ov004_020b2444.c
+ *   [44] 0x020b2574  src/func_ov004_020b2574.c
+ *   [45] 0x020b265c  src/func_ov004_020b265c.c
+ *   [46] 0x020b27f4  src/func_ov004_020b27f4.c
+ *   [47] 0x020b2880  src/func_ov004_020b2880.c
+ *   [48] 0x020b290c  src/func_ov004_020b290c.c
+ *   [49] 0x020b2980  src/func_ov004_020b2980.c
+ *   [50] 0x020b298c  src/func_ov004_020b298c.c
+ *   [51] 0x020b2990  src/func_ov004_020b2990.c
+ *   [52] 0x020b2994  src/func_ov004_020b2994.c
+ *   [53] 0x020b299c  src/func_ov004_020b299c.c
+ *   [54] 0x020b29a0  src/func_ov004_020b29a0.cpp
+ *   [55] 0x020b29c0  src/_ZN11dScMgBase_cD2Ev.cpp
+ *   [56] 0x020b2a18  src/_ZN11dScMgBase_cD0Ev.cpp
+ *   [57] 0x020b2a84  src/_ZN11dScMgBase_cD1Ev.cpp
+ */
+
+/* Includes: union of the legacy files', first-seen in ROM-ascending
+ * processing order. NOT verified for header ordering constraints (e.g. a
+ * common.h-before-X rule) -- watch for new compile errors after this. */
+#include "types.h"
+#include "private/ov004_obj_vtbl.h"
+#include "decl_common.h"
+#include "dScMgBase_c.h"
+
+/* Local shadow declarations carried from the legacy files verbatim.
+ * NOT reconciled against real project headers -- check include/*.h for
+ * each of these before compiling; a real header should usually win. */
+/* shadow struct 'E' */
+struct E
+{
+  char pad[0x20];
+  int f20;
+  char pad2[0x134 - 0x24];
+};
+
+/* shadow struct 'Vec2' */
+struct Vec2 {
+    short x;
+    short y;
+};
+
+/* shadow struct 'Obj004' */
+struct Obj004 {
+    char pad0[0x10];
+    struct Vec2 pos;  /* 0x10 */
+    char pad14[0xc];
+    int f20;          /* 0x20 */
+    char pad24[0xe];
+    unsigned short f32;/* 0x32 */
+    char pad34[0x100];
+};
+
+/* shadow struct 'Elem' */
+struct Elem { char pad[0x134]; };
+
+/* shadow typedef 'struct' (renamed Elem -> IconEnt: collides with
+   func_ov004_020b0cac's own 0x134-stride `struct Elem`) */
+typedef struct { char pad[0x1a]; short field_1a; char pad2[4]; } IconEnt;
+
+/* shadow typedef 'int' */
+typedef int s32;
+
+/* shadow typedef 'long' */
+typedef long long s64;
+
+/* shadow struct 'Base' */
+struct Base {
+    virtual void v0(); virtual void v1(); virtual void v2(); virtual void v3();
+    virtual void v4(); virtual void v5(); virtual void v6(); virtual void v7();
+    virtual void v8(); virtual void v9(); virtual void v10(); virtual void v11();
+    virtual void v12(); virtual void v13(); virtual void v14(); virtual void v15();
+    virtual void v16(); virtual void v17(); virtual void v18(void*);
+};
+
+#define AT(p,off) ((int*)(int)((char*)(p)+(off)))
+
+extern "C" {
+/* reconciled survivors -- see the CONFLICT notes below */
+extern char data_0209b308[];
+extern char *data_ov004_020bbfa8[];
+extern unsigned char data_020a0de8[];
+extern void func_ov004_020b87e0(void* a, void* b);
+extern void func_ov004_020b3194(void *c);
+extern unsigned char data_020a0e40;
+extern unsigned char data_020a0de9[][4];
+extern unsigned char data_020a0dea[][4];
+extern unsigned char data_020a0deb[][4];
+extern int data_ov004_020bc150;
+extern void func_ov004_020b422c(struct Obj004 *o);
+extern int func_ov004_020b40c0(struct Obj004 *o);
+extern void func_ov004_020b3278(struct Elem *e, int c, int a1, int a2, int a3, int s1, short s2);
+extern void FreeGfxSlotsById(int);
+extern char data_ov004_020bf3e8[];
+extern s32 GetGameLanguage(void);
+extern void Hud_RenderSprite(void *fn, s32 a, s32 b, s32 r3arg, s32 stack_arg);
+extern "C" void func_02012790(int);
+extern "C" void RenderOamMainScreen(int a0, int a1, int a2, int a3, int a4);
+extern int func_ov004_020ad6f4(void* self, int a);
+extern int func_ov004_020adc3c(void* self);
+extern int func_02013580(int a, int b);
+extern void func_ov004_020b1710(int a0, int x, int y, int val);
+extern void func_ov004_020b2444(int a1, int a2, int num, int a4, int a5, int sel, int idx);
+extern void func_ov004_020b1b78(void *a0, int a1);
+extern void func_ov004_020b1ba0(void *a0, s32 a1);
+extern void func_ov004_020b1de8(int r0, int r1, int r2, int r3);
+extern void RenderOamBothScreens(void* a0, int a1, int a2, int a3, int a4, int a5);
+extern void func_ov004_020afc18(void* a0, int a1, int a2, int a3, int a4, int a5);
+extern void func_ov004_020b1d60(void *a0, int a1, int a2, int a3, int a4);
+extern void func_ov004_020b0e84(struct Obj *, int);
+extern void DrawOamSprite(int a, int b, int c, int d);
+extern int data_ov006_02137cd8[];
+extern void func_ov004_020af948(void *a, int b, int c, void *m);
+extern void _ZN2GX15SetGraphicsModeEiii(int a, int b, int c);
+extern void _ZN3GXS15SetGraphicsModeEi(int a);
+extern void func_ov004_020b2980(void);
+extern void func_ov004_020b290c(void);
+extern void _ZN2GX12SetBankForBGEt(u16 a);
+extern void _ZN2GX13SetBankForOBJEt(u16 a);
+extern void _ZN2GX15SetBankForSubBGEt(u16 a);
+extern void _ZN2GX16SetBankForSubOBJEt(u16 a);
+extern void *func_ov004_020adc68(int id);
+extern void DecompressLZ16(void *src, void *dst);
+extern void Ov004_Deallocate(void *p);
+extern void _ZN4CP1527FlushAndInvalidateDataCacheEjj(void *p, u32 len);
+extern void _ZN2GX11LoadOBJPlttEPKvjj(const void *p, u32 a, u32 b);
+extern void _ZN3GXS11LoadOBJPlttEPKvjj(const void *p, u32 a, u32 b);
+extern void func_ov004_020b0d30(void);
+extern int data_ov004_020beb6c;
+extern u8 data_0209d45c;
+extern int data_ov004_020bbfe4[];
+extern int data_ov004_020beb74[];
+extern u8 data_0209d454;
+extern void **data_0209d4a8;
+extern int data_0208ee44;
+extern unsigned int LoadCompressedFileAt(int fileID, void *target);
+extern unsigned int _ZN3G2S13GetBG1CharPtrEv(void);
+extern void *_ZN3G2S12GetBG1ScrPtrEv(void);
+extern void _ZN2GX15DisableAllBanksEv(void);
+extern void *func_ov004_020b929c(void *);
+/* TUBUILD CONFLICT -- alternate declaration of data_ov004_020beb68, from the legacy file for func_ov004_020b0b1c, NOT applied: extern int data_ov004_020beb68; */
+/* TUBUILD CONFLICT -- alternate declaration of GetGameLanguage, from the legacy file for func_ov004_020b0e84, NOT applied: extern "C" int GetGameLanguage(void); */
+/* TUBUILD CONFLICT -- alternate declaration of GetGameLanguage, from the legacy file for func_ov004_020b1710, NOT applied: extern int GetGameLanguage(void); */
+/* TUBUILD CONFLICT -- alternate declaration of func_ov004_020b1ea4, from the legacy file for func_ov004_020b1710, NOT applied: extern void func_ov004_020b1ea4(int x, int a1, int val, int a3, int a4, int mode, int fb); */
+/* TUBUILD CONFLICT -- alternate declaration of Hud_RenderSprite, from the legacy file for func_ov004_020b1710, NOT applied: extern void Hud_RenderSprite(void *a0, int a1, int a2, int a3, int a4); */
+/* TUBUILD CONFLICT -- alternate declaration of data_ov004_020beb68, from the legacy file for func_ov004_020b1aec, NOT applied: extern int *data_ov004_020beb68; */
+/* TUBUILD CONFLICT -- alternate declaration of data_ov004_020beb68, from the legacy file for func_ov004_020b1b08, NOT applied: extern void *data_ov004_020beb68; */
+/* TUBUILD CONFLICT -- alternate declaration of data_ov004_020beb68, from the legacy file for func_ov004_020b1b40, NOT applied: extern void *data_ov004_020beb68; */
+/* TUBUILD CONFLICT -- alternate declaration of func_ov004_020b2444, from the legacy file for func_ov004_020b1bc8, NOT applied: extern void func_ov004_020b2444(int a, int b, int c, int d, int e, int f, int g); */
+/* TUBUILD CONFLICT -- alternate declaration of func_ov004_020b1ea4, from the legacy file for func_ov004_020b1bc8, NOT applied: extern void func_ov004_020b1ea4(int a, int b, int c, int d, int e, int f, int g); */
+/* TUBUILD CONFLICT -- alternate declaration of GetGameLanguage, from the legacy file for func_ov004_020b1de8, NOT applied: extern int GetGameLanguage(void); */
+/* TUBUILD CONFLICT -- alternate declaration of func_ov004_020b2444, from the legacy file for func_ov004_020b1e44, NOT applied: extern void func_ov004_020b2444(int a, int b, int c, int d, int e, int f, int g); */
+/* TUBUILD CONFLICT -- alternate declaration of GetGameLanguage, from the legacy file for func_ov004_020b1e44, NOT applied: extern int GetGameLanguage(void); */
+/* TUBUILD CONFLICT -- alternate declaration of func_ov004_020b1d60, from the legacy file for func_ov004_020b1ea4, NOT applied: extern void func_ov004_020b1d60(void *glyph, int x, int a2, int a3, int a4); */
+/* TUBUILD CONFLICT -- alternate declaration of func_ov004_020b1d60, from the legacy file for func_ov004_020b2444, NOT applied: extern void func_ov004_020b1d60(int a0, int a1, int a2, int a3, int a4); */
+/* TUBUILD CONFLICT -- alternate declaration of GetGameLanguage, from the legacy file for func_ov004_020b2574, NOT applied: extern int GetGameLanguage(void); */
+/* TUBUILD CONFLICT -- alternate declaration of DrawOamSprite, from the legacy file for func_ov004_020b2574, NOT applied: extern void DrawOamSprite(void *arg0, void *arg1, int arg2, void *arg3); */
+/* TUBUILD CONFLICT -- alternate declaration of FreeGfxSlotsById, from the legacy file for func_ov004_020b265c, NOT applied: extern void FreeGfxSlotsById(int arg); */
+/* TUBUILD CONFLICT -- alternate declaration of GetGameLanguage, from the legacy file for func_ov004_020b27f4, NOT applied: extern int GetGameLanguage(void); */
+/* TUBUILD CONFLICT -- alternate declaration of GetGameLanguage, from the legacy file for func_ov004_020b2880, NOT applied: extern int GetGameLanguage(void); */
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 57 -- _ZN11dScMgBase_cD1Ev, 0x020b2a84, size 0x58 */
+/* -------------------------------------------------------------------------- */
+// @symbol _ZN11dScMgBase_cD1Ev
+/* ~dScMgBase_c() is declared, NOT defined inline, in dScMgBase_c.h -- see
+   the header's own note. mwcc doesn't inline a body this size into a
+   descendant (measured against dScMgD3DBase_c), so this is a REAL
+   out-of-line definition, same shape Stage.h uses for a leaf -- except
+   here D0Ev.cpp carries an identical copy for its own key-function TU. */
+dScMgBase_c::~dScMgBase_c()
+{
+    data_ov004_020beb68 = 0;
+    func_ov004_020b929c((char *)this + 0xf4);
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 56 -- _ZN11dScMgBase_cD0Ev, 0x020b2a18, size 0x6c */
+/* -------------------------------------------------------------------------- */
+/* NO SEPARATE DEFINITION. In the merged TU the one ~dScMgBase_c() below emits
+   the whole D1/D0/D2 variant group; a second definition of the same destructor
+   is ill-formed. dScMgBase_c.h's inline `operator delete` is what lets mwcc
+   emit D0 at all (see include/Actor.h's D0 route note). */
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 55 -- _ZN11dScMgBase_cD2Ev, 0x020b29c0, size 0x58 */
+/* -------------------------------------------------------------------------- */
+/* NO SEPARATE DEFINITION -- emitted by the ~dScMgBase_c() above. The legacy
+   free-function form hand-wrote what mwcc generates for D2 on its own: the
+   two vtable stores and the `bl _ZN9ActorBaseD2Ev`. Writing it out here as
+   well would define the symbol twice. */
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 54 -- func_ov004_020b29a0, 0x020b29a0, size 0x20 */
+/* -------------------------------------------------------------------------- */
+extern "C" void func_ov004_020b29a0(Base *c, void *arg) { c->v18(arg); }
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 53 -- func_ov004_020b299c, 0x020b299c, size 0x4 */
+/* -------------------------------------------------------------------------- */
+extern "C" {
+// @symbol func_ov004_020b299c
+/* recovered: renamed to Class_Method */
+void func_ov004_020b299c(void)
+{
+}
+
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 52 -- func_ov004_020b2994, 0x020b2994, size 0x8 */
+/* -------------------------------------------------------------------------- */
+extern "C" {
+// @symbol func_ov004_020b2994
+// recovered name: dScMgBase_c_OnTurnIntoEgg
+/* recovered: renamed to Class_Method */
+/* dScMgBase_c::OnTurnIntoEgg - recovered from vtable slot identity */
+int func_ov004_020b2994(void)
+{
+    return 1;
+}
+
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 51 -- func_ov004_020b2990, 0x020b2990, size 0x4 */
+/* -------------------------------------------------------------------------- */
+extern "C" {
+// @symbol func_ov004_020b2990
+// recovered name: dScMgBase_c_Virtual50
+/* recovered: renamed to Class_Method */
+/* dScMgBase_c::Virtual50 - recovered from vtable slot identity */
+void func_ov004_020b2990(void)
+{
+}
+
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 50 -- func_ov004_020b298c, 0x020b298c, size 0x4 */
+/* -------------------------------------------------------------------------- */
+extern "C" {
+// @symbol func_ov004_020b298c
+// recovered name: dScMgBase_c_OnGroundPounded
+/* recovered: renamed to Class_Method */
+/* dScMgBase_c::OnGroundPounded - recovered from vtable slot identity */
+void func_ov004_020b298c(void)
+{
+}
+
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 49 -- func_ov004_020b2980, 0x020b2980, size 0xc */
+/* -------------------------------------------------------------------------- */
+extern "C" {
+void func_ov004_020b2980(void) { _ZN2GX15DisableAllBanksEv(); }
+
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 48 -- func_ov004_020b290c, 0x020b290c, size 0x74 */
+/* -------------------------------------------------------------------------- */
+extern "C" {
+void SetBg0Offset(int a, int b);
+void SetBg1Offset(int a, int b);
+void SetBg2Offset(int a, int b);
+void SetBg3Offset(int a, int b);
+void SetSubBg0Offset(int a, int b);
+void SetSubBg1Offset(int a, int b);
+void SetSubBg2Offset(int a, int b);
+void SetSubBg3Offset(int a, int b);
+
+void func_ov004_020b290c(void) {
+    SetBg0Offset(0, 0);
+    SetBg1Offset(0, 0);
+    SetBg2Offset(0, 0);
+    SetBg3Offset(0, 0);
+    SetSubBg0Offset(0, 0);
+    SetSubBg1Offset(0, 0);
+    SetSubBg2Offset(0, 0);
+    SetSubBg3Offset(0, 0);
+}
+
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 47 -- func_ov004_020b2880, 0x020b2880, size 0x8c */
+/* -------------------------------------------------------------------------- */
+extern "C" {
+// @symbol func_ov004_020b2880
+// recovered name: dScMgBase_c_Kill
+/* recovered: renamed to Class_Method, declarations from a shared header */
+/* recovered: renamed to Class_Method */
+/* dScMgBase_c::Kill - recovered from vtable slot identity */
+void func_ov004_020b2880(void)
+{
+    int f;
+    *(volatile unsigned short *)0x400100a = (*(volatile unsigned short *)0x400100a & 0x43) | 0x10;
+    *(volatile unsigned short *)0x400100a = *(volatile unsigned short *)0x400100a & ~0x40;
+    *(volatile unsigned short *)0x400100a = *(volatile unsigned short *)0x400100a & ~3;
+    SetSubBg1Offset(0, 0);
+    data_0209d454 &= ~2;
+    f = GetGameLanguage();
+    LoadCompressedFileAt(data_ov004_020bc00c[f], (void *)_ZN3G2S13GetBG1CharPtrEv());
+    LoadCompressedFileAt(0x5b, _ZN3G2S12GetBG1ScrPtrEv());
+}
+
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 46 -- func_ov004_020b27f4, 0x020b27f4, size 0x8c */
+/* -------------------------------------------------------------------------- */
+extern "C" {
+// @symbol func_ov004_020b27f4
+// recovered name: dScMgBase_c_AfterClsn
+/* recovered: renamed to Class_Method, declarations from a shared header */
+/* recovered: renamed to Class_Method */
+/* dScMgBase_c::AfterClsn - recovered from vtable slot identity */
+void func_ov004_020b27f4(void)
+{
+    int f;
+    *(volatile unsigned short *)0x400000a = *(volatile unsigned short *)0x400000a & ~3;
+    *(volatile unsigned short *)0x400000a = (*(volatile unsigned short *)0x400000a & 0x43) | 0x1000;
+    *(volatile unsigned short *)0x400000a = *(volatile unsigned short *)0x400000a & ~0x40;
+    SetBg1Offset(0, 0);
+    data_0209d45c &= ~2;
+    f = GetGameLanguage();
+    LoadCompressedFileAt(data_ov004_020bbff8[f], (void *)func_02054ea8());
+    LoadCompressedFileAt(0x67, _ZN2G212GetBG1ScrPtrEv());
+}
+
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 45 -- func_ov004_020b265c, 0x020b265c, size 0x198 */
+/* -------------------------------------------------------------------------- */
+extern "C" {
+void func_ov004_020b265c(char *obj)
+{
+    void *p;
+
+    data_ov004_020beb6c = 0;
+    obj[0x68] = 0;
+    *(int *)(obj + 0x6c) = -1;
+    *(vu32 *)0x4001000u |= 0x10000u;
+    _ZN2GX15SetGraphicsModeEiii(1, 0, 0);
+    _ZN3GXS15SetGraphicsModeEi(0);
+    *(vu32 *)0x4001000u &= 0xffcfffefu;
+    *(vu16 *)0x4000304u |= 0x8000u;
+    func_ov004_020b2980();
+    func_ov004_020b290c();
+    *(vu32 *)0x4000000u &= ~0x7000000u;
+    *(vu32 *)0x4000000u &= ~0x38000000u;
+    _ZN2GX12SetBankForBGEt(3);
+    _ZN2GX13SetBankForOBJEt(0x10);
+    data_0209d45c = 0x10;
+    _ZN2GX15SetBankForSubBGEt(4);
+    _ZN2GX16SetBankForSubOBJEt(8);
+    p = func_ov004_020adc68(data_ov004_020bbfe4[GetGameLanguage()]);
+    {
+        char *dst = (char *)0x6400000; dst += 0x4000;
+        DecompressLZ16(p, dst);
+    }
+    {
+        char *dst = (char *)0x6600000; dst += 0x4000;
+        DecompressLZ16(p, dst);
+    }
+    Ov004_Deallocate(p);
+    p = func_ov004_020adc68(0xc3);
+    _ZN4CP1527FlushAndInvalidateDataCacheEjj(p, 0x100u);
+    _ZN2GX11LoadOBJPlttEPKvjj(p, 0x100u, 0x100u);
+    _ZN3GXS11LoadOBJPlttEPKvjj(p, 0x100u, 0x100u);
+    Ov004_Deallocate(p);
+    data_0209d454 = 0x10;
+    data_ov004_020beb74[1] = (int)obj;
+    data_0209d4a8 = (void **)data_ov004_020beb74;
+    func_ov004_020b0d30();
+    FreeGfxSlotsById(0x1d);
+    data_0208ee44 = 1;
+}
+
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 44 -- func_ov004_020b2574, 0x020b2574, size 0xe8 */
+/* -------------------------------------------------------------------------- */
+extern "C" {
+void func_ov004_020b2574(int arg0, int arg1)
+{
+    void **table = (void **)data_ov004_020bbfa8[GetGameLanguage()];
+    void *sb = table[0];
+    int b = 12;
+    if (arg1 != 1) {
+        DrawOamSprite((int)sb, b, b, 0);
+        DrawOamSprite((int)((void **)data_ov004_020bbfa8[GetGameLanguage()])[1], 30, b, 0);
+        func_ov004_020b2444(0x30, b, arg0, 1, -1, 2, 0);
+    } else {
+        int i;
+        for (i = 0; i < arg0; i++) {
+            func_ov004_020af948(sb, b, 12, 0);
+            b += 16;
+        }
+    }
+}
+
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 43 -- func_ov004_020b2444, 0x020b2444, size 0x130 */
+/* -------------------------------------------------------------------------- */
+extern "C" {
+void func_ov004_020b2444(int a1, int a2, int num, int a4, int a5, int sel, int idx)
+{
+    unsigned t = num;
+    int digits = 0;
+    int off;
+    if (num != 0) {
+        do {
+            t /= 10;
+            digits++;
+        } while (t != 0);
+    }
+    off = 0;
+    switch (sel) {
+    case 0:
+        off = (digits << 3) - 8;
+        break;
+    case 1:
+        break;
+    case 2:
+        off = (digits << 4) - 0x10;
+        break;
+    }
+    if (num == 0) {
+        func_ov004_020b1d60((void *)data_ov006_02137cd8[idx], a1, a2, a4, a5);
+        return;
+    }
+    if (num <= 0)
+        return;
+    do {
+        func_ov004_020b1d60((void *)data_ov006_02137cd8[num % 10 + idx], a1 + off, a2, a4, a5);
+        num /= 10;
+        off -= 0x10;
+    } while (num > 0);
+}
+
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 41 -- func_ov004_020b1ea4, 0x020b1ea4, size 0x37c */
+/* -------------------------------------------------------------------------- */
+extern "C" {
+void func_ov004_020b1ea4(int x, int a1, int val, int a3, int a4, int mode, int fb) {
+    int d[6];
+    int i;
+    if (val >= 0xf423f) val = 0xf423f;
+    for (i = 0; i < 6; i++) d[i] = 0;
+    d[0] = val;
+    while (d[0] >= 0x186a0) { d[5]++; d[0] -= 0x186a0; }
+    while (d[0] >= 0x2710) { d[4]++; d[0] -= 0x2710; }
+    while (d[0] >= 0x3e8) { d[3]++; d[0] -= 0x3e8; }
+    while (d[0] >= 0x64) { d[2]++; d[0] -= 0x64; }
+    while (d[0] >= 0xa) { d[1]++; d[0] -= 0xa; }
+    if (fb == 0) fb = 0xa;
+    else if (fb == 0) fb = 0xa;
+    else if (fb == 0x14) fb = 0x1e;
+    else if (fb == 0xa) fb = 0xa;
+
+    if (d[5] != 0) {
+        int r5;
+        int k;
+        if (mode == 0) x -= 0x14;
+        else if (mode == 1) x -= 0x28;
+        r5 = 5;
+        for (k = 0; k < 5; k++) {
+            func_ov004_020b1d60((void*)data_ov006_02137cd8[fb + d[r5]], x, a1, a3, a4);
+            x += 8;
+            r5--;
+        }
+    } else if (d[4] != 0) {
+        int r5;
+        int k;
+        if (mode == 0) x -= 0x10;
+        else if (mode == 1) x -= 0x20;
+        r5 = 4;
+        for (k = 0; k < 4; k++) {
+            func_ov004_020b1d60((void*)data_ov006_02137cd8[fb + d[r5]], x, a1, a3, a4);
+            x += 8;
+            r5--;
+        }
+    } else if (d[3] != 0) {
+        int r5;
+        int k;
+        if (mode == 0) x -= 0xc;
+        else if (mode == 1) x -= 0x18;
+        r5 = 3;
+        for (k = 0; k < 3; k++) {
+            func_ov004_020b1d60((void*)data_ov006_02137cd8[fb + d[r5]], x, a1, a3, a4);
+            x += 8;
+            r5--;
+        }
+    } else if (d[2] != 0) {
+        int r5;
+        int k;
+        if (mode == 0) x -= 8;
+        else if (mode == 1) x -= 0x10;
+        r5 = 2;
+        for (k = 0; k < 2; k++) {
+            func_ov004_020b1d60((void*)data_ov006_02137cd8[fb + d[r5]], x, a1, a3, a4);
+            x += 8;
+            r5--;
+        }
+    } else if (d[1] != 0) {
+        if (mode == 0) x -= 4;
+        else if (mode == 1) x -= 8;
+        func_ov004_020b1d60((void*)data_ov006_02137cd8[fb + d[1]], x, a1, a3, a4);
+        x += 8;
+    }
+    func_ov004_020b1d60((void*)data_ov006_02137cd8[d[0] + fb], x, a1, a3, a4);
+}
+
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 40 -- func_ov004_020b1e44, 0x020b1e44, size 0x60 */
+/* -------------------------------------------------------------------------- */
+extern "C" {
+void func_ov004_020b1e44(int arg) {
+    int r;
+    func_ov004_020b2444(0x80, 0xc, arg, 1, -1, 0, 0);
+    r = GetGameLanguage();
+    DrawOamSprite(((int *)data_ov004_020bbfa8[r])[3], 0x60, 0xc, 0);
+}
+
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 39 -- func_ov004_020b1e34, 0x020b1e34, size 0x10 */
+/* -------------------------------------------------------------------------- */
+extern "C" {
+/* func_ov004_020b1e34 @ 0x20b1e34 (ov004) -- veneer: ldr r1,[r0,#0xb4]; b func_ov004_020b0e84. */
+void func_ov004_020b1e34(void* a) {
+    func_ov004_020b0e84((struct Obj *)a, (int)*(u32*)((char*)a + 0xb4));
+}
+
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 38 -- func_ov004_020b1de8, 0x020b1de8, size 0x4c */
+/* -------------------------------------------------------------------------- */
+extern "C" {
+void func_ov004_020b1de8(int r0, int r1, int r2, int r3) {
+    int idx = GetGameLanguage();
+    int *t = (int *)data_ov004_020bbfa8[idx];
+    func_ov004_020b1d60((void *)t[2], r0, r1, r2, r3);
+}
+
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 37 -- func_ov004_020b1d60, 0x020b1d60, size 0x88 */
+/* -------------------------------------------------------------------------- */
+extern "C" {
+void func_ov004_020b1d60(void* a0, int a1, int a2, int a3, int a4){
+  if(a3==-1){
+    RenderOamBothScreens(a0, a1, a2, -1, a4, 0);
+    return;
+  }
+  if(a3==0){
+    func_ov004_020af868(a0, a1, a2, -1, a4, 0);
+    return;
+  }
+  func_ov004_020afc18(a0, a1, a2, -1, a4, 0);
+}
+
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 36 -- func_ov004_020b1cf0, 0x020b1cf0, size 0x70 */
+/* -------------------------------------------------------------------------- */
+extern "C" {
+void func_ov004_020afdd0(int a, int b, int c, int d, int e);
+void func_ov004_020b1cf0(int a, int b, int c, int sel, int e){
+  if(sel == -1){ func_ov004_020afdd0(a,b,c,-1,e); return; }
+  if(sel == 0){ Hud_RenderSprite((void *)a,b,c,-1,e); return; }
+  RenderOamMainScreen(a,b,c,-1,e);
+}
+
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 35 -- func_ov004_020b1c68, 0x020b1c68, size 0x88 */
+/* -------------------------------------------------------------------------- */
+extern "C" {
+void func_ov004_020b1c68(void* a0, int a1, int a2, int a3, int a4, int a5){
+  if(a3==-1){
+    RenderOamBothScreens(a0, a1, a2, -1, a4, a5);
+    return;
+  }
+  if(a3==0){
+    func_ov004_020af868(a0, a1, a2, -1, a4, (void *)a5);
+    return;
+  }
+  func_ov004_020afc18(a0, a1, a2, -1, a4, a5);
+}
+
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 34 -- func_ov004_020b1bc8, 0x020b1bc8, size 0xa0 */
+/* -------------------------------------------------------------------------- */
+extern "C" {
+void func_ov004_020b1bc8(char *a0, int a1, int a2, int a3)
+{
+    func_ov004_020b1de8(a1, a2, a3, -1);
+    int v = *(int*)(a0+0xac);
+    if (v < 0x64) {
+        func_ov004_020b2444(a1+0x12, a2, v, a3, -1, 2, 0);
+    } else {
+        func_ov004_020b1ea4(a1+0xc, a2, v, a3, -1, 2, 0x32);
+    }
+}
+
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 33 -- func_ov004_020b1ba0, 0x020b1ba0, size 0x28 */
+/* -------------------------------------------------------------------------- */
+extern "C" {
+void func_ov004_020b1ba0(void *c, s32 delta)
+{
+    s32 v = *(s32*)((char*)c + 0xa8) + delta;
+    if (v < 0) v = 0;
+    else if (v > 0x270f) v = 0x270f;
+    *(s32*)((char*)c + 0xa8) = v;
+}
+
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 32 -- func_ov004_020b1b78, 0x020b1b78, size 0x28 */
+/* -------------------------------------------------------------------------- */
+extern "C" {
+void func_ov004_020b1b78(void* c, int val) {
+    int r2 = *(int*)((char*)c + 0xa8);
+    r2 -= val;
+    if (r2 < 0) r2 = 0;
+    else if (r2 > 0x270f) r2 = 0x270f;
+    *(int*)((char*)c + 0xa8) = r2;
+}
+
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 31 -- func_ov004_020b1b40, 0x020b1b40, size 0x38 */
+/* -------------------------------------------------------------------------- */
+extern "C" {
+void func_ov004_020b1b40(void *c) {
+    void *v = data_ov004_020beb68;
+    if (v == 0) return;
+    func_ov004_020b1ba0(v, (s32)c);
+}
+
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 30 -- func_ov004_020b1b08, 0x020b1b08, size 0x38 */
+/* -------------------------------------------------------------------------- */
+extern "C" {
+void func_ov004_020b1b08(void *c) {
+    void *v = data_ov004_020beb68;
+    if (v == 0) return;
+    func_ov004_020b1b78(v, (int)c);
+}
+
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 29 -- func_ov004_020b1aec, 0x020b1aec, size 0x1c */
+/* -------------------------------------------------------------------------- */
+extern "C" {
+int func_ov004_020b1aec(void) {
+    int *g = (int *)data_ov004_020beb68;
+    if (g != 0) return *(int *)((char *)g + 0xbc);
+    return 0;
+}
+
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 28 -- func_ov004_020b1a5c, 0x020b1a5c, size 0x90 */
+/* -------------------------------------------------------------------------- */
+extern "C" {
+/* func_ov004_020b1a5c at 0x020b1a5c
+ *
+ * Matched byte-for-byte with mwccarm 1.2/sp2p3 (ov004).
+ */
+s32 GetGameLanguage(void);
+extern int data_ov004_020b944c[];
+extern char* data_ov004_020bbfa8[];
+void func_ov004_020b1a5c(int a0, int a1){
+  int lim = data_ov004_020b944c[a1];
+  int c = a0;
+  int r4 = 0xf4;
+  int i;
+  if (c >= lim) c = lim;
+  func_ov004_020b1ea4(0xf4, 0xc, c, 1, -1, 1, 0xa);
+  r4 = r4 - (0x44 - (6 - a1) * 8);
+  i = GetGameLanguage();
+  RenderOamMainScreen(*(int*)(data_ov004_020bbfa8[i]+0x10), r4, 0xc, -1, -1);
+}
+
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 27 -- func_ov004_020b19f0, 0x020b19f0, size 0x6c */
+/* -------------------------------------------------------------------------- */
+extern "C" {
+s32 GetGameLanguage(void);
+extern char* data_ov004_020bbfa8[];
+int func_ov004_020b19f0(void* self){
+  int i = GetGameLanguage();
+  RenderOamMainScreen(*(int*)(data_ov004_020bbfa8[i]+0x20), 0x20, 0xc, -1, -1);
+  func_ov004_020b1ea4(0x48, 0xc, (int)self, 1, -1, 2, 0xa);
+}
+
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 26 -- func_ov004_020b1710, 0x020b1710, size 0x2e0 */
+/* -------------------------------------------------------------------------- */
+extern "C" {
+void func_ov004_020b1710(int a0, int x, int y, int val)
+{
+    int idx = GetGameLanguage();
+    int mode = ((int *)data_0209b308)[2];
+    int *tbl = (int *)data_ov004_020bbfa8[idx];
+    int flag = 0;
+
+    if (mode == 5) {
+        int major;
+        int rem;
+        int pct;
+        if (val < 0)
+            return;
+        major = val / 60;
+        func_ov004_020b1ea4(x - 0x18, y, major, flag, -1, 1, 0xa);
+        Hud_RenderSprite((void *)tbl[7], x - 0x10, y, -1, -1);
+
+        rem = val % 60;
+        pct = rem * 100;
+        func_ov004_020b1ea4(x - 8, y, pct / 600, 0, -1, 1, 0xa);
+
+        rem = val % 60;
+        pct = rem * 100 / 60;
+        func_ov004_020b1ea4(x, y, pct % 10, 0, -1, 1, 0xa);
+        return;
+    }
+
+    if (mode == 3)
+        flag = 1;
+    switch (mode) {
+    case 1:
+        Hud_RenderSprite((void *)tbl[9], 0x58, y, -1, -1);
+        Hud_RenderSprite((void *)tbl[1], 0x68, y, -1, -1);
+        break;
+    case 2:
+        Hud_RenderSprite((void *)tbl[2], 0x58, y, -1, -1);
+        Hud_RenderSprite((void *)tbl[1], 0x68, y, -1, -1);
+        break;
+    case 3:
+        flag = 1;
+        break;
+    case 4:
+        Hud_RenderSprite((void *)tbl[0x13], 0x58, y, -1, -1);
+        Hud_RenderSprite((void *)tbl[1], 0x68, y, -1, -1);
+        break;
+    default:
+        break;
+    }
+
+    if (flag == 1)
+        x -= 4;
+    if (val < 0)
+        return;
+    if (flag == 1)
+        func_ov004_020b1ea4(x, y, val, 0, -1, 1, 0xa);
+    else
+        func_ov004_020b2444(x, y, val, 0, -1, 1, 0);
+}
+
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 25 -- func_ov004_020b14f0, 0x020b14f0, size 0x220 */
+/* -------------------------------------------------------------------------- */
+extern "C" {
+void func_ov004_020b14f0(char* self)
+{
+    int a2 = 0x28;
+    int n;
+    int i;
+    int j;
+    int k;
+
+    *AT(self, 0x60) += 1;
+    if (*(int*)(self + 0x60) >= 0x8c)
+        *(int*)(self + 0x60) = 0x64;
+
+    n = func_ov004_020ad6f4(self, *(int*)(self + 0x4658));
+    if (n < 0)
+        goto neg;
+
+    i = 0;
+    if (n > 0)
+    {
+        do
+        {
+            if (*(int*)(self + 0x60) >= i * 0x14)
+                func_ov004_020b1710((int)self, 0xa8, a2, func_02013580(func_ov004_020adc3c(self), i));
+            a2 += 0x18;
+            i++;
+        } while (i < n);
+    }
+
+    if (*(int*)(self + 0x60) >= n * 0x14)
+    {
+        if (*(int*)(self + 0x60) < 0x64 || *(int*)(self + 0x60) - 0x64 < 0x14)
+            func_ov004_020b1710((int)self, 0xa8, a2, *(int*)(self + 0x4658));
+        else
+            func_ov004_020b1710((int)self, 0xa8, a2, -1);
+    }
+
+    a2 += 0x18;
+    j = n + 1;
+    if (j >= 5)
+        return;
+    do
+    {
+        if (*(int*)(self + 0x60) >= j * 0x14)
+            func_ov004_020b1710((int)self, 0xa8, a2, func_02013580(func_ov004_020adc3c(self), j));
+        a2 += 0x18;
+        j++;
+    } while (j < 5);
+    return;
+
+neg:
+    k = 0;
+    do
+    {
+        if (*(int*)(self + 0x60) >= k * 0x14)
+            func_ov004_020b1710((int)self, 0xa8, a2, func_02013580(func_ov004_020adc3c(self), k));
+        a2 += 0x18;
+        k++;
+    } while (k < 5);
+
+    if (*(int*)(self + 0x60) < 0x64)
+        return;
+
+    if (*(int*)(self + 0x60) - 0x64 < 0x14)
+        func_ov004_020b1710((int)self, 0xa8, a2 + 0x10, *(int*)(self + 0x4658));
+    else
+        func_ov004_020b1710((int)self, 0xa8, a2 + 0x10, -1);
+}
+
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 24 -- func_ov004_020b0e84, 0x020b0e84, size 0x66c */
+/* -------------------------------------------------------------------------- */
+static inline int sdiv4(int d) {
+    int t = d + (int)((unsigned)(d >> 1) >> 0x1e);
+    return t >> 2;
+}
+
+static inline int sdiv2(int e) {
+    return (e + (int)((unsigned)e >> 31)) >> 1;
+}
+
+extern "C" void func_ov004_020b0e84(Obj *sl, int sb) {
+    void *r8;
+    int r7;
+    int i;
+    int pin;
+
+    {
+        int t = sl->unk50;
+        if (t < sb) {
+            sl->unk54 = 1;
+            sl->unk50 = sb;
+            func_02012790(0x1bb);
+        } else if (t > sb) {
+            sl->unk54 = 0;
+            sl->unk50 = sb;
+        } else if (sl->unk54 > 0) {
+            if (sl->f68() == 2) {
+                sl->unk54 += 2;
+            } else {
+                sl->unk54 += 1;
+            }
+        }
+    }
+
+    r8 = (void *)data_ov004_020bbfa8[GetGameLanguage()];
+    r7 = 0xa8;
+
+    if (sb < 5) {
+        r7 -= 8;
+        i = 0;
+        if (sb > 0) {
+            do {
+                if (i == sb - 1 && sl->unk54 > 0 && sl->unk54 % 0x3c >= 0x1e) {
+                    break;
+                }
+                RenderOamMainScreen(((int *)r8)[0x24 / 4], 0xf0, r7, -1, -1);
+                i += 1;
+                r7 -= 0x10;
+            } while (i < sb);
+        }
+        if (sl->unk54 >= sl->unk58) {
+            sl->unk54 = 0;
+        }
+        return;
+    }
+
+    if (sb == 5) {
+        int u = sl->unk54;
+        if (u == 0) {
+            func_ov004_020b1ea4(0xf0, r7, 5, 1, 0, 0, 0x14);
+            RenderOamMainScreen(((int *)r8)[0x2c / 4], 0xf0, r7, -1, -1);
+        } else if (u < sl->unk58) {
+            r7 -= 8;
+            i = 0;
+            do {
+                RenderOamMainScreen(((int *)r8)[0x24 / 4], 0xf0, r7, -1, -1);
+                i += 1;
+                r7 -= 0x10;
+            } while (i < 4);
+            if (sl->unk54 % 0x3c < 0x1e) {
+                RenderOamMainScreen(((int *)r8)[0x24 / 4], 0xf0, r7, -1, -1);
+            }
+        } else {
+            int d = (sl->unk58 + 0x20) - sl->unk54;
+            r7 -= sdiv4(d);
+            i = 0;
+            do {
+                RenderOamMainScreen(((int *)r8)[0x24 / 4], 0xf0, r7, -1, -1);
+                {
+                    unsigned e = (sl->unk58 + 0x20) - sl->unk54;
+                    r7 -= sdiv2((int)e);
+                }
+                i += 1;
+            } while (i < 5);
+        }
+        if (sl->f68() == 2) {
+            int b = sl->unk58;
+            int c = sl->unk54;
+            if (c == b || c == b + 1) {
+                func_02012790(0x1ba);
+            }
+        } else if (sl->unk54 == sl->unk58) {
+            func_02012790(0x1ba);
+        }
+        if (sl->unk54 >= sl->unk58 + 0x20) {
+            sl->unk54 = 0;
+        }
+        return;
+    }
+
+    {
+        int rem = sb % 5;
+        if (rem == 0) {
+            int u = sl->unk54;
+            if (u == 0) {
+                func_ov004_020b1ea4(0xf0, r7, sb, 1, 0, 0, 0x14);
+                RenderOamMainScreen(((int *)r8)[0x2c / 4], 0xf0, r7, -1, -1);
+            } else if (u < sl->unk58) {
+                func_ov004_020b1ea4(0xf0, r7, sb - 5, 1, 0, 0, 0x14);
+                RenderOamMainScreen(((int *)r8)[0x2c / 4], 0xf0, r7, -1, -1);
+                r7 -= 0x18;
+                i = 0;
+                do {
+                    if (i == 4 && sl->unk54 > 0 && sl->unk54 % 0x3c >= 0x1e) {
+                        break;
+                    }
+                    RenderOamMainScreen(((int *)r8)[0x24 / 4], 0xf0, r7, -1, -1);
+                    i += 1;
+                    r7 -= 0x10;
+                } while (i < 5);
+            } else {
+                int d;
+                func_ov004_020b1ea4(0xf0, r7, sb - 5, 1, 0, 0, 0x14);
+                RenderOamMainScreen(((int *)r8)[0x2c / 4], 0xf0, r7, -1, -1);
+                d = (sl->unk58 + 0x20) - sl->unk54;
+                r7 -= sdiv4(d);
+                i = 0;
+                do {
+                    unsigned e = (sl->unk58 + 0x20) - sl->unk54;
+                    r7 -= sdiv2((int)e);
+                    RenderOamMainScreen(((int *)r8)[0x24 / 4], 0xf0, r7, -1, -1);
+                    i += 1;
+                } while (i < 5);
+            }
+            if (sl->f68() == 2) {
+                int b = sl->unk58;
+                int c = sl->unk54;
+                if (c == b || c == b + 1) {
+                    func_02012790(0x1ba);
+                }
+            } else if (sl->unk54 == sl->unk58) {
+                func_02012790(0x1ba);
+            }
+            if (sl->unk54 >= sl->unk58 + 0x20) {
+                sl->unk54 = 0;
+            }
+            return;
+        }
+
+        {
+            int base5 = (sb / 5) * 5;
+            func_ov004_020b1ea4(0xf0, r7, base5, 1, 0, 0, 0x14);
+        }
+        RenderOamMainScreen(((int *)r8)[0x2c / 4], 0xf0, r7, -1, -1);
+        r7 -= 0x18;
+        sb = 0;
+        if (rem > 0) {
+            pin = rem - 1;
+            do {
+                if (sb == pin && sl->unk54 > 0 && sl->unk54 % 0x3c >= 0x1e) {
+                    break;
+                }
+                RenderOamMainScreen(((int *)r8)[0x24 / 4], 0xf0, r7, -1, -1);
+                sb += 1;
+                r7 -= 0x10;
+            } while (sb < rem);
+        }
+        if (sl->unk54 >= sl->unk58) {
+            sl->unk54 = 0;
+        }
+    }
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 23 -- func_ov004_020b0de0, 0x020b0de0, size 0xa4 */
+/* -------------------------------------------------------------------------- */
+extern "C" {
+s32 GetGameLanguage(void);
+extern char* data_ov004_020bbfa8[];
+
+#define LAU(p) ((int)(p))
+
+void func_ov004_020b0de0(void* c_)   /* decl_common.h types this void * */
+{
+    char* c = (char*)c_;
+    if (*(unsigned char*)(c + 0xc3) == 0)
+        return;
+    if ((unsigned int)*(unsigned char*)(c + 0xc4) < 4U) {
+        unsigned short* p = (unsigned short*)LAU(c + 0xc0);
+        *p = (unsigned short)(*p + 1);
+        if ((unsigned int)*(unsigned short*)(c + 0xc0) >= 0x30U) {
+            *(unsigned short*)(c + 0xc0) = 0U;
+            {
+                unsigned char* q = (unsigned char*)LAU(c + 0xc4);
+                *q = (unsigned char)(*q + 1);
+            }
+        }
+        if ((unsigned int)*(unsigned short*)(c + 0xc0) >= 0x18U)
+            return;
+    }
+    {
+        int i = GetGameLanguage();
+        RenderOamMainScreen(*(int*)(data_ov004_020bbfa8[i] + 0x28), 0xc0, 0xb0, -1, -1);
+    }
+}
+
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 22 -- func_ov004_020b0d8c, 0x020b0d8c, size 0x54 */
+/* -------------------------------------------------------------------------- */
+extern "C" {
+void func_ov004_020b0d8c(void *c, s32 arg1, s32 arg2)
+{
+    s32 idx1 = *(s32*)((char*)c + 0x5c);
+    s32 r4 = (idx1 < 0x14) ? 0x11 : 0x12;
+    s32 idx2 = GetGameLanguage();
+    void **table = (void**)data_ov004_020bbfa8[idx2];
+    void *fn = table[r4];
+    Hud_RenderSprite(fn, arg1, arg2, -1, -1);
+}
+
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 21 -- func_ov004_020b0d30, 0x020b0d30, size 0x5c */
+/* -------------------------------------------------------------------------- */
+extern "C" {
+void func_ov004_020b0d30(void) {
+    FreeGfxSlotsById(0x1d);
+    int i = 0;
+    do {
+        ((IconEnt *)data_ov004_020bebe8)[i].field_1a = 0;
+        i++;
+    } while (i < 0x40);
+    int j = 0;
+    char* p = data_ov004_020bf3e8;
+    do {
+        j++;
+        *p++ = 0;
+    } while (j < 3);
+}
+
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 20 -- func_ov004_020b0cac, 0x020b0cac, size 0x84 */
+/* -------------------------------------------------------------------------- */
+extern "C" {
+void func_ov004_020b0cac(int c, int a1, int a2, int a3, int arg5, short arg6) {
+    int i;
+    for (i = 0; i < 3; i++) {
+        if (*(int*)((char*)&((struct Elem *)data_ov004_020bf648)[i] + 0x20) == 0x1d) {
+            func_ov004_020b3278(&((struct Elem *)data_ov004_020bf648)[i], c, a1, a2, a3, arg5, arg6);
+            return;
+        }
+    }
+}
+
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 19 -- func_ov004_020b0b1c, 0x020b0b1c, size 0x190 */
+/* -------------------------------------------------------------------------- */
+extern "C" {
+int func_ov004_020b0b1c(int arg) {
+    int idx = data_020a0e40;
+    int flag = 0;
+    int i;
+
+    if (data_020a0de8[idx * 4] != 0) {
+        if (data_020a0de9[idx][0] != 0) flag = 1;
+    }
+
+    if (flag) {
+        struct Obj004 *o = (struct Obj004 *)data_ov004_020bf648;
+        int cx = data_020a0dea[idx][0];
+        int cy = data_020a0deb[idx][0];
+        for (i = 0; i < 3; i++, o++) {
+            if (arg == o->f20 && o->f32 != 0) {
+                volatile struct Vec2 pos;
+                int dx, dy;
+                short ty;
+                pos.x = o->pos.x;
+                ty = o->pos.y;
+                dx = pos.x - cx;
+                pos.y = ty;
+                dy = pos.y - cy;
+                if (dx <= 0x40 && dx >= -0x40 && dy <= 0x13 && dy >= -0x13) {
+                    if (data_ov004_020bc150 != 0) {
+                        func_ov004_020b422c(o);
+                        if ((unsigned int)(arg - 1) <= 1) {
+                            *(int *)((int)data_ov004_020beb68 + 0x4000 + 0x630) = 1;
+                        }
+                    }
+                }
+            }
+        }
+    } else {
+        int j;
+        struct Obj004 *o = (struct Obj004 *)data_ov004_020bf648;
+        for (j = 0; j < 3; j++, o++) {
+            if (arg == o->f20) {
+                if (func_ov004_020b40c0(o) != 0) {
+                    if (arg == 1) {
+                    } else if (arg == 2) {
+                    } else {
+                        data_ov004_020bc150 = 1;
+                    }
+                    return 1;
+                }
+            }
+        }
+    }
+    return 0;
+}
+
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 18 -- FreeGfxSlotsById, 0x020b0aa0, size 0x7c */
+/* -------------------------------------------------------------------------- */
+extern "C" {
+void FreeGfxSlotsById(int arg)
+{
+  if (arg == 0x1d)
+  {
+    int i;
+    struct E *p;
+    p = (struct E *)data_ov004_020bf648;
+    for (i = 0; i < 3; i++)
+    {
+      if (p->f20 != 0x1d)
+      {
+        func_ov004_020b3194(p);
+      }
+      p++;
+    }
+
+  }
+  else
+  {
+    int i;
+    struct E *p;
+    p = (struct E *)data_ov004_020bf648;
+    for (i = 0; i < 3; i++)
+    {
+      if (arg == p->f20)
+      {
+        func_ov004_020b3194(p);
+      }
+      p++;
+    }
+
+  }
+}
+
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 17 -- func_ov004_020b0a54, 0x020b0a54, size 0x4c */
+/* -------------------------------------------------------------------------- */
+extern "C" {
+void func_ov004_020b0a54(int c_)     /* decl_common.h types this int */
+{
+    void* c = (void*)c_;
+    void* g = data_ov004_020beb68;
+    if (!g) return;
+    func_ov004_020b87e0((char*)g + 0xcc, c);
+    *(int*)((char*)data_ov004_020beb68 + 0x60) = 0;
+}
+
+}
+
+/* -------------------------------------------------------------------------- */
+/* ROM ordinal 16 -- func_ov004_020b0a38, 0x020b0a38, size 0x1c */
+/* -------------------------------------------------------------------------- */
+extern "C" {
+void func_ov004_020b0a38(void)
+{
+    int *p = (int *)data_ov004_020beb68;
+    if (p == 0) return;
+    *(int *)((char *)p + 0xf0) = 0;
+}
+
+}
